Extract Ghostty environment bootstrap from cmuxApp.swift
cmuxApp.swift is over the file-length hard cap, so CI rejects any net
growth; the handlesExternalEvents fix added 5 lines. Move the
self-contained Ghostty environment setup into
cmuxApp+GhosttyEnvironment.swift, shrinking cmuxApp.swift by 87 lines.
